requirements.txt
alembic==1.3.2
autopep8==1.4.4
cffi==1.13.2
Click==7.0
Flask==1.1.1
Flask-Login==0.4.1
Flask-Migrate==2.5.2
Flask-Misaka==1.0.0
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
itsdangerous==1.1.0
Jinja2==2.10.3
Mako==1.1.0
MarkupSafe==1.1.1
misaka==2.1.1
Pillow==7.0.0
pycodestyle==2.5.0
pycparser==2.19
python-dateutil==2.8.1
python-dotenv==0.10.3
python-editor==1.0.4
six==1.13.0
SQLAlchemy==1.3.12
Werkzeug==0.16.0
WTForms==2.2.1
